requirements.txt
numpy
pypdf2
python-dotenv
requests
sentence-transformers
streamlit
tqdm